+++ src/main/java/com/atlauncher/gui/dialogs/instancesettings/JavaInstanceSettingsTab.java
@@ -42,6 +42,8 @@ import javax.swing.JTextField;
 import javax.swing.SpinnerNumberModel;
 import javax.swing.SwingUtilities;
 import javax.swing.border.Border;
+import javax.swing.event.ChangeEvent;
+import javax.swing.event.ChangeListener;
 
 import com.atlauncher.App;
 import com.atlauncher.builders.HTMLBuilder;
@@ -121,6 +123,16 @@ public class JavaInstanceSettingsTab extends JPanel {
         initialMemoryModel.setMaximum((systemRam == 0 ? null : systemRam));
         initialMemory = new JSpinner(initialMemoryModel);
         ((JSpinner.DefaultEditor) initialMemory.getEditor()).getTextField().setColumns(5);
+        initialMemory.addChangeListener(new ChangeListener() {
+            @Override
+            public void stateChanged(ChangeEvent e) {
+                JSpinner s = (JSpinner) e.getSource();
+                // if initial memory is larger than maximum memory, make initial memory match
+                if ((Integer) s.getValue() > (Integer) maximumMemory.getValue()) {
+                    initialMemory.setValue(maximumMemory.getValue());
+                }
+            }
+        });
         add(initialMemory, gbc);
 
         // Maximum Memory Settings
@@ -149,6 +161,16 @@ public class JavaInstanceSettingsTab extends JPanel {
         maximumMemoryModel.setMaximum((systemRam == 0 ? null : systemRam));
         maximumMemory = new JSpinner(maximumMemoryModel);
         ((JSpinner.DefaultEditor) maximumMemory.getEditor()).getTextField().setColumns(5);
+        maximumMemory.addChangeListener(new ChangeListener() {
+            @Override
+            public void stateChanged(ChangeEvent e) {
+                JSpinner s = (JSpinner) e.getSource();
+                // if initial memory is larger than maximum memory, make initial memory match
+                if ((Integer) initialMemory.getValue() > (Integer) s.getValue()) {
+                    initialMemory.setValue(s.getValue());
+                }
+            }
+        });
         add(maximumMemory, gbc);
 
         // Perm Gen Settings